Write listener for initial system notifications
Modify Notification.addListener('press', ...) to only register an on click listener. To
determine if the application was initially opened (not in background)
via a notification click create a new listener
Notification.addListener('initial',...) to handle this event.

Add a new callback to addListener which is called if the app is loaded
by the user and not by clicking a notification